Add GResource helper class
